refactor(rust-api): extract character handlers into characters.rs (sc-1636)
Move the 17 contiguous character endpoint handlers (list/create/get/
update/archive/purge character, reference + look + lora sub-resource
handlers, create_character_test_job) out of lib.rs into a dedicated
characters.rs child module.

Same pattern as auth.rs: `use super::*;` in the module sees the crate
root (AppState, the Character* DTOs, project_store input types, ApiError,
shared helpers) unchanged; handlers are pub(crate) and glob re-imported
at the root so the router wiring in create_app stays unqualified.
Behavior-preserving. lib.rs 8,253 -> 7,889 lines.
